requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.14
pygame==2.5.2
pystray==0.19.5
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
requests==2.31.0
pytest==8.1.1
